Proper File Structures for themes
