+++ styles/theme/colors.ts
@@ -1,0 +1,35 @@
+// Extend the theme to include custom colors
+const colors = {
+    heading: {
+        light: '#3e02ab',
+        dark: '#21c54b'
+    },
+    strong: {
+        light: '#ab0202',
+        dark: '#ffb256'
+    },
+    em: {
+        light: '#8600d4',
+        dark: '#d895ff'
+    },
+    code: {
+        color: {
+            light: '#e01e5a',
+            dark: '#ffb256'
+        },
+        bg: {
+            light: 'rgba(29, 28, 29, 0.04)',
+            dark: 'rgb(255 255 255 / 9%)'
+        }
+    },
+    link: {
+        color: { light: '#1d612e', dark: '#3ec4c1' },
+        hover: {
+            light: '#1d612e',
+            dark: '#42f1f4'
+        },
+        active: 'rgba(48, 140, 122, 0.3)'
+    }
+};
+
+export { colors };
