+++ styles/theme/index.ts
@@ -1,82 +1,19 @@
 import { extendTheme, ThemeOverride, ColorMode } from '@chakra-ui/react';
-import { createBreakpoints } from '@chakra-ui/theme-tools';
 import { mdx } from './mdx-styles';
-
-// 2. Extend the theme to include custom colors, fonts, etc
-const colors = {
-    heading: {
-        light: '#3e02ab',
-        dark: '#21c54b'
-    },
-    strong: {
-        light: '#ab0202',
-        dark: '#ffb256'
-    },
-    em: {
-        light: '#8600d4',
-        dark: '#d895ff'
-    },
-    code: {
-        color: {
-            light: '#e01e5a',
-            dark: '#ffb256'
-        },
-        bg: {
-            light: 'rgba(29, 28, 29, 0.04)',
-            dark: 'rgb(255 255 255 / 9%)'
-        }
-    },
-    link: {
-        color: { light: '#1d612e', dark: '#3ec4c1' },
-        hover: {
-            light: '#1d612e',
-            dark: '#42f1f4'
-        },
-        active: 'rgba(48, 140, 122, 0.3)'
-    }
-};
-
-const breaks = [350, 450, 650, 960, 1200, 1440];
-
-const mq = breaks.map((bp) => `@media (min-width: ${bp}px)`);
-
-export const bpDesktopOnly = `@media (min-width: ${1200 + 1}px)`;
-
-const BREAKPOINTS = {
-    xs: '350px',
-    sm: '450px',
-    md: '650px',
-    lg: '960px',
-    xl: '1200px',
-    '2xl': '1440px'
-};
-
-const breakpoints = createBreakpoints(BREAKPOINTS);
+import { colors } from './colors';
+import { breakpoints } from './breakpoints';
+import { general, fonts } from './styles';
 
 const config = {
     initialColorMode: 'light' as ColorMode,
     useSystemColorMode: false
 };
 
-const general = {
-    link: {
-        fontWeight: 'semibold',
-        transition: 'color 0.15s',
-        transitionTimingFunction: 'ease-out',
-        _hover: {
-            borderBottom: '1px solid'
-        }
-    }
-};
-
 const overrides: ThemeOverride & { mdx: Record<string, unknown> } & {
     general: Record<string, unknown>;
 } = {
     colors,
-    fonts: {
-        body: 'Roboto, sans-serif, system-ui',
-        heading: 'Roboto, sans-serif, system-ui'
-    },
+    fonts,
     breakpoints,
     config,
     mdx,
@@ -84,7 +21,5 @@ const overrides: ThemeOverride & { mdx: Record<string, unknown> } & {
 };
 
 const theme = extendTheme(overrides);
-
-export { BREAKPOINTS, mq };
 
 export default theme;
